requirements.txt
torch==2.2.0
torchvision==0.17.0
flask==2.3.2
pillow==10.0.0
numpy==2.0.0 